refactor(cli): organize the dev command
