+++ packages/cli/src/commands/dev/consts.ts
@@ -1,0 +1,1 @@
+export const TUNNEL_URL = 'https://novu.sh/api/tunnels';
