+++ packages/cli/src/commands/dev/enums.ts
@@ -1,0 +1,11 @@
+export enum CloudRegionEnum {
+  US = 'us',
+  EU = 'eu',
+  STAGING = 'staging',
+}
+
+export enum DashboardUrlEnum {
+  US = 'https://dashboard.novu.co',
+  EU = 'https://eu.dashboard.novu.co',
+  STAGING = 'https://dev.dashboard.novu.co',
+}
