+++ packages/cli/src/commands/dev/index.ts
@@ -1,0 +1,2 @@
+export { devCommand } from './dev';
+export { DevCommandOptions } from './types';
