+++ packages/cli/src/commands/dev/types.ts
@@ -1,0 +1,15 @@
+import { CloudRegionEnum } from './enums';
+
+export type DevCommandOptions = {
+  port: string;
+  origin: string;
+  region: `${CloudRegionEnum}`;
+  studioPort: string;
+  dashboardUrl: string;
+  route: string;
+};
+
+export type LocalTunnelResponse = {
+  id: string;
+  url: string;
+};
